partial implementation of common file fragment classifier
git-svn-id: https://showcase.fhstp.ac.at/republic/mmc/trunk@753 76e0c08c-8e14-4548-9514-92ecd803f47c

+++ collating/fragment/include/fragment_classifier_p.h
@@ -8,7 +8,7 @@ typedef struct
 {
     void* mSoHandler;
     struct _FragmentClassifier* mFcHandler;
-    float mWeight;
+    int mWeight;
     fc_new_ptr mFcNew;
     fc_classify_ptr mFcClassify;
     fc_free_ptr mFcFree;
